+++ CHANGES.md
@@ -1,5 +1,7 @@
 ### dev
 
+### 0.5.0 / 2022-04-08
+
 * Refactored Biggs::Concern to use Biggs::Extractor
 * Removed activerecord dependency and use only activesupport instead
 * Renamed Biggs::ActiveRecordAdapter to Biggs::Concern, include with 'include Biggs'